关键词开发 zdl

+++ src/main/java/com/cn/zbx/controller/KeyWordsController.java
@@ -19,6 +19,11 @@ import com.cn.zbx.pojo.KeyWords;
 import com.cn.zbx.service.IKeyWordsService;
 import com.cn.zbx.vo.KeyWordsVO;
 
+/**
+ * 关键词Controller
+ * @author zdl
+ *
+ */
 @Controller
 @RequestMapping(value="/keyWords")
 public class KeyWordsController {
@@ -26,6 +31,12 @@ public class KeyWordsController {
 	@Autowired
 	IKeyWordsService keyWordsService;
 	
+	/**
+	 * 关键词主页面初始化方法
+	 * @param request
+	 * @param response
+	 * @return
+	 */
 	@ResponseBody
 	@RequestMapping(value="/initKeyWordsMain", method = { RequestMethod.GET, RequestMethod.POST })
 	public String initKeyWordsMain(HttpServletRequest request, HttpServletResponse response){
@@ -56,6 +67,13 @@ public class KeyWordsController {
 		return JSONObject.toJSONString(resultMap);
 	}
 	
+	/**
+	 * 根据条件查询关键词信息
+	 * @param request
+	 * @param response
+	 * @param keyWords
+	 * @return
+	 */
 	@ResponseBody
 	@RequestMapping(value="/selectKeyWordsByParam", method = { RequestMethod.GET, RequestMethod.POST })
 	public String selectKeyWordsByParam(HttpServletRequest request, HttpServletResponse response, KeyWords keyWords){
@@ -81,6 +99,12 @@ public class KeyWordsController {
 		return JSONObject.toJSONString(resultMap);
 	}
 	
+	/**
+	 * 根据id删除关键词（若该关键词关联有文章或者视频不能被删除）
+	 * @param request
+	 * @param response
+	 * @return
+	 */
 	@ResponseBody
 	@RequestMapping(value="/deleteById", method = { RequestMethod.GET, RequestMethod.POST })
 	public String deleteById(HttpServletRequest request, HttpServletResponse response){
@@ -88,17 +112,41 @@ public class KeyWordsController {
 		
 		String keyWordsId = request.getParameter("keyWordsId");
 		if(keyWordsId == null || "".equals(keyWordsId)){
+			resultMap.put("errorMsg", "数据传输错误！");
 			resultMap.put("success", false);
+			return JSONObject.toJSONString(resultMap);
 		}
-		int result = keyWordsService.deleteByPrimaryKey(Integer.valueOf(keyWordsId));
-		if(result <= 0){
-			resultMap.put("success", false);
+		KeyWords keyWordsParam = new KeyWords();
+		keyWordsParam.setId(Integer.valueOf(keyWordsId));
+		List<KeyWordsVO> keyWordsList = keyWordsService.selectBySelectParam(keyWordsParam);
+		if(keyWordsList != null && keyWordsList.size() != 0){
+			KeyWordsVO tKeyWordsVO = keyWordsList.get(0);
+			if(tKeyWordsVO.getArticleCount() == 0 && tKeyWordsVO.getVideoCount() == 0){
+				int result = keyWordsService.deleteByPrimaryKey(Integer.valueOf(keyWordsId));
+				if(result <= 0){
+					resultMap.put("errorMsg", "数据提交错误！");
+					resultMap.put("success", false);
+				} else {
+					resultMap.put("success", true);
+				}
+			} else {
+				resultMap.put("errorMsg", "该关键词关联有文章或者视频不能被删除！");
+				resultMap.put("success", false);
+			}
 		} else {
-			resultMap.put("success", true);
+			resultMap.put("errorMsg", "查询关键词失败！");
+			resultMap.put("success", false);
 		}
 		return JSONObject.toJSONString(resultMap);
 	}
 	
+	/**
+	 * 添加关键词信息（添加的关键词名字已有的不能被添加）
+	 * @param request
+	 * @param response
+	 * @param keyWords
+	 * @return
+	 */
 	@ResponseBody
 	@RequestMapping(value="/addKeyWords", method = { RequestMethod.GET, RequestMethod.POST })
 	public String addtClassify(HttpServletRequest request, HttpServletResponse response, KeyWords keyWords){
@@ -108,25 +156,42 @@ public class KeyWordsController {
 		KeyWords keyWordsParam = new KeyWords();
 		if(keyWords.getName() != null && !"".equals(keyWords.getName())){
 			keyWordsParam.setName(keyWords.getName());
-		}
-		if(keyWords.getExcerpt() != null && !"".equals(keyWords.getExcerpt())){
-			keyWordsParam.setExcerpt(keyWords.getExcerpt());
-		}
-		if(keyWords.getState() != null && !"".equals(keyWords.getState())){
-			keyWordsParam.setState(keyWords.getState());
-		}
-		keyWordsParam.setMakedate(currentDate);
-		keyWordsParam.setModifydate(currentDate);
-		
-		int result = keyWordsService.insertSelective(keyWordsParam);
-		if(result <= 0){
-			resultMap.put("success", false);
+			int num = keyWordsService.selectCountBySelectParam(keyWordsParam);
+			if(num > 0){
+				resultMap.put("errorMsg", "已有该关键词，不能被添加！");
+				resultMap.put("success", false);
+			} else {
+				if(keyWords.getExcerpt() != null && !"".equals(keyWords.getExcerpt())){
+					keyWordsParam.setExcerpt(keyWords.getExcerpt());
+				}
+				if(keyWords.getState() != null && !"".equals(keyWords.getState())){
+					keyWordsParam.setState(keyWords.getState());
+				}
+				keyWordsParam.setMakedate(currentDate);
+				keyWordsParam.setModifydate(currentDate);
+				
+				int result = keyWordsService.insertSelective(keyWordsParam);
+				if(result <= 0){
+					resultMap.put("errorMsg", "数据提交错误！");
+					resultMap.put("success", false);
+				} else {
+					resultMap.put("success", true);
+				}
+			}
 		} else {
-			resultMap.put("success", true);
+			resultMap.put("errorMsg", "数据传输错误！");
+			resultMap.put("success", false);
 		}
 		return JSONObject.toJSONString(resultMap);
 	}
 	
+	/**
+	 * 修改关键词信息（修改的关键词名字已有的不能被修改）
+	 * @param request
+	 * @param response
+	 * @param keyWords
+	 * @return
+	 */
 	@ResponseBody
 	@RequestMapping(value="/updateSelectById", method = { RequestMethod.GET, RequestMethod.POST })
 	public String updateSelectById(HttpServletRequest request, HttpServletResponse response, KeyWords keyWords){
@@ -134,23 +199,43 @@ public class KeyWordsController {
 		
 		KeyWords keyWordsParam = new KeyWords();
 		if(keyWords.getId() != null && !"".equals(keyWords.getId())){
-			keyWordsParam.setId(keyWords.getId());
-		}
-		if(keyWords.getName() != null && !"".equals(keyWords.getName())){
-			keyWordsParam.setName(keyWords.getName());
-		}
-		if(keyWords.getExcerpt() != null && !"".equals(keyWords.getExcerpt())){
-			keyWordsParam.setExcerpt(keyWords.getExcerpt());
-		}
-		if(keyWords.getState() != null && !"".equals(keyWords.getState())){
-			keyWordsParam.setState(keyWords.getState());
-		}
-		keyWordsParam.setModifydate(new Date());
-		int result = keyWordsService.updateByPrimaryKeySelective(keyWordsParam);
-		if(result <= 0){
-			resultMap.put("success", false);
-		} else {
-			resultMap.put("success", true);
+			KeyWords tKeyWords = keyWordsService.selectByPrimaryKey(keyWords.getId());
+			if(tKeyWords != null){
+				boolean b = true;
+				if(keyWords.getName() != null && !"".equals(keyWords.getName())){
+					keyWordsParam.setName(keyWords.getName());
+					List<KeyWordsVO> keyWordsList = keyWordsService.selectBySelectParam(keyWordsParam);
+					for(KeyWordsVO kv : keyWordsList){
+						if(kv.getId() != tKeyWords.getId() && kv.getName().equals(keyWords.getName())){
+							b = false;
+							break;
+						}
+					}
+				}
+				if(b){
+					keyWordsParam.setId(keyWords.getId());
+					if(keyWords.getExcerpt() != null && !"".equals(keyWords.getExcerpt())){
+						keyWordsParam.setExcerpt(keyWords.getExcerpt());
+					}
+					if(keyWords.getState() != null && !"".equals(keyWords.getState())){
+						keyWordsParam.setState(keyWords.getState());
+					}
+					keyWordsParam.setModifydate(new Date());
+					int result = keyWordsService.updateByPrimaryKeySelective(keyWordsParam);
+					if(result <= 0){
+						resultMap.put("errorMsg", "数据提交失败！");
+						resultMap.put("success", false);
+					} else {
+						resultMap.put("success", true);
+					}
+				} else {
+					resultMap.put("errorMsg", "修改的关键词名字已存在！");
+					resultMap.put("success", false);
+				} 
+			} else {
+				resultMap.put("errorMsg", "分类查询失败！");
+				resultMap.put("success", false);
+			}
 		}
 		return JSONObject.toJSONString(resultMap);
 	}
